ctime - lastIncoming がラップラウンドしないように int64_t に拡張した。
ctime を取った後に lastIncoming が更新されて ctime < lastIncoming になった場合に、差が巨大になって 3600 秒を超える判定になり、プログラムが終了していたのを修正。

+++ core/common/servmgr.cpp
@@ -1880,7 +1880,7 @@ int ServMgr::idleProc(ThreadInfo *thread)
         if (servMgr->isRoot)
         {
             // 1時間着信がなかったら終了する。…なぜ？
-            if ((servMgr->lastIncoming) && ((ctime-servMgr->lastIncoming) > 60*60))
+            if ((servMgr->lastIncoming) && (((int64_t)ctime - servMgr->lastIncoming) > 60*60))
             {
                 peercastInst->saveSettings();
                 sys->exit();
